requirements.txt
flask
flask-smorest 
python-dotenv
sqlalchemy
flask-sqlalchemy
rq
redis
requests
flask-jwt-extended
passlib
